requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
django-grappelli==2.12.1
django-ckeditor==5.6.1

Pillow==5.3.0